+++ Source/ACE.Server/Program.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Diagnostics;
 using System.Globalization;
 using System.IO;
 using System.Reflection;
@@ -123,10 +122,7 @@ namespace ACE.Server
             if (IsRunningInContainer)
                 log.Info("ACEmulator is running in a container...");
 
-            var assembly = Assembly.GetExecutingAssembly();
-            var fileVersionInfo = FileVersionInfo.GetVersionInfo(assembly.Location);
-            var serverVersion = fileVersionInfo.ProductVersion;
-            Console.Title = @$"ACEmulator - v{serverVersion}";
+            Console.Title = @$"ACEmulator - v{ServerBuildInfo.FullVersion}";
 
             var configFile = Path.Combine(exeLocation, "Config.js");
             var configConfigContainer = Path.Combine(containerConfigDirectory, "Config.js");
